refactor: extract getting application SDK to utility function
